Fix build after 3debf6297591ac4d638e498e0a08093e2e45844e
Because of missing imports and some undefined variables the latest code wouldn't compile.

I have fixed it.

The only thing I am not sure about is dispatchConfig: PropTypes.bool.isRequired
Not sure if it's truly bool or if this is valid, but it works finally again, but with proper log timezone!